Introduce new RngBitGenerator HLO
The new instruction has the same signature as xla::BitGeneratorTy
what takes a key and a state and returns uniformly distributed
random bits and a new value for the state.

Its aim is to enable backend specific lowering for the various random
bit generator algorithms what should unlock optimization opportunities.

PiperOrigin-RevId: 293569472
Change-Id: I4f69d4f9858378fb1241435032ef75657933c056