Fix TF-TRT control edges between TRTEngineOp nodes
CPed from upstream PR #37294

Ignore control out edges from const node.

When we loop throug the input edges, only non-const input edges are
added to the list of connections. With this fix we do the same for
the output edges, otherwise the EngineInfo structures could be
inconsistent if we have multiple TRT segments.